feat(slugs): expose isSlug, addContextToSlug, and removeContextFromSlug functions
also remove circular dependencies

AFFECTS PACKAGES:
@esri/hub-common
@esri/hub-content
